判断wxCode有无上传

+++ src/main/java/com/scoprion/mall/wx/service/delivery/WxDeliveryServiceImpl.java
@@ -84,8 +84,17 @@ public class WxDeliveryServiceImpl implements WxDeliveryService {
     @Override
     public BaseResult deleteDelivery(Long id, String wxCode) {
         String userId = WxUtil.getOpenId(wxCode);
+        if(StringUtils.isEmpty(userId)){
+            return BaseResult.parameterError();
+        }
         //获取收货地址是否是默认地址
         Delivery delivery=wxDeliveryMapper.findById(id);
+        if(StringUtils.isEmpty(id.toString())){
+            return BaseResult.parameterError();
+        }
+        if (delivery == null){
+            return BaseResult.notFound();
+        }
         if(CommonEnum.DEFAULT_ADDRESS.getCode().equals(delivery.getDefaultAddress())){
             Integer result = wxDeliveryMapper.deleteDelivery(id);
             if (result <= 0) {
